+++ packages/article-in-depth/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.92.6](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.92.5...@times-components/article-in-depth@3.92.6) (2024-12-03)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.92.5](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.92.4...@times-components/article-in-depth@3.92.5) (2024-12-02)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
 ## [3.92.4](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.92.3...@times-components/article-in-depth@3.92.4) (2024-11-22)
 
 **Note:** Version bump only for package @times-components/article-in-depth