+++ packages/article-magazine-comment/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.87.6](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.87.5...@times-components/article-magazine-comment@3.87.6) (2024-12-03)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [3.87.5](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.87.4...@times-components/article-magazine-comment@3.87.5) (2024-12-02)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
 ## [3.87.4](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.87.3...@times-components/article-magazine-comment@3.87.4) (2024-11-22)
 
 **Note:** Version bump only for package @times-components/article-magazine-comment